+++ src/components/textarea/editForm/TextArea.edit.display.js
@@ -1,3 +1,6 @@
+import _ from 'lodash';
+import Formio from '../../../Formio';
+
 export default [
   {
     key: 'inputMask',
@@ -32,6 +35,107 @@ export default [
       ]
     },
     weight: 415
+  },
+  {
+    type: 'checkbox',
+    input: true,
+    key: 'isUploadEnabled',
+    label: 'Enable Image Upload',
+    weight: 415.1,
+    conditional: {
+      json: {
+        or: [
+          {
+            '===': [
+              { var: 'data.editor' },
+              'quill'
+            ]
+          },
+          {
+            '==': [
+              { var: 'data.editor' },
+              ''
+            ]
+          }
+        ]
+      }
+    }
+  },
+  {
+    type: 'select',
+    input: true,
+    key: 'uploadStorage',
+    label: 'Storage',
+    placeholder: 'Select your file storage provider',
+    weight: 415.2,
+    tooltip: 'Which storage to save the files in.',
+    valueProperty: 'value',
+    dataSrc: 'custom',
+    data: {
+      custom() {
+        return _.map(Formio.providers.storage, (storage, key) => ({
+          label: storage.title,
+          value: key
+        }));
+      }
+    },
+    conditional: {
+      json: {
+        '===': [
+          { var: 'data.isUploadEnabled' },
+          true
+        ]
+      }
+    }
+  },
+  {
+    type: 'textfield',
+    input: true,
+    key: 'uploadUrl',
+    label: 'Url',
+    weight: 415.3,
+    placeholder: 'Enter the url to post the files to.',
+    tooltip: 'See <a href=\'https://github.com/danialfarid/ng-file-upload#server-side\' target=\'_blank\'>https://github.com/danialfarid/ng-file-upload#server-side</a> for how to set up the server.',
+    conditional: {
+      json: { '===': [{ var: 'data.uploadStorage' }, 'url'] }
+    }
+  },
+  {
+    type: 'textarea',
+    key: 'uploadOptions',
+    label: 'Custom request options',
+    tooltip: 'Pass your custom xhr options(optional)',
+    rows: 5,
+    editor: 'ace',
+    input: true,
+    weight: 415.4,
+    placeholder: `{
+      "withCredentials": true
+    }`,
+    conditional: {
+      json: {
+        '===': [{
+          var: 'data.uploadStorage'
+        }, 'url']
+      }
+    }
+  },
+  {
+    type: 'textfield',
+    input: true,
+    key: 'uploadDir',
+    label: 'Directory',
+    placeholder: '(optional) Enter a directory for the files',
+    tooltip: 'This will place all the files uploaded in this field in the directory',
+    weight: 415.5,
+    conditional: {
+      json: {
+        '===': [
+          { var: 'data.isUploadEnabled' },
+          true
+        ]
+      }
+    }
   },
   {
     type: 'select',