+++ Changelog.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ## [Unreleased]
 ### Added
  - reCAPTCHA component
+ - Ability to upload images to File Storage for Quill WYSIWYG
 
 ## 3.11.1
 ### Fixed